requirements.txt
praw==7.6.0
click==8.1.3
python-dotenv==0.20.0